Add support for Morello to clang, llvm, lld, lldb, libcxx, libcxxabi and compiler-rt.
Known limitations:
  - Only the small code model is supported for the pure capability ABI
  - ELF is the only supported binary format
  - Only emulated TLS is supported in the pure capability ABI
  - No support for sanitizers
  - Dynamic linking in the pure capability ABI is supported but untested
  - IFUNC support for the pure capability ABI is untested
  - The ptrace interface support in lldb likely still needs to be modified
    to match what the kernel implements
  - lldb is missing support for decoding capabilities and will only print
    the hex value

Several people have contributed to this patch:
 Alexandros Lamprineas
 Amilendra Kodithuwakku - Morello lld support
 Arnaud de Grandmaison
 Diana Picus (diana.picus@linaro.org) - Morello lldb support
 Javed Absar
 John Brawn
 Peter Smith (peter.smith@linaro.org) - Original lld implementation
 Petr Pavlu - Generic lldb capability support, DWARF
 Silviu Baranga

+++ compiler-rt/lib/sanitizer_common/sanitizer_allocator.cpp
@@ -29,7 +29,7 @@ const char *SecondaryAllocatorName = "LargeMmapAllocator";
 #define SANITIZER_USE_MALLOC
 #endif
 
-static constexpr usize InternalDefaultAlign = Max((usize)8, sizeof(void*));
+static constexpr usize InternalDefaultAlign = Max((usize)8, (usize)sizeof(void*));
 
 // ThreadSanitizer for Go uses libc malloc/free.
 #if SANITIZER_GO || defined(SANITIZER_USE_MALLOC)
